Backward connections in one CC learn by themselves
But there is a problem: First, these backward connections don't learn correctly. (all of them had been decreasing) The learning rate of + and - are split and the + one has been set to the higher rate.
Currently, after a while, they learn correctly. However, if learning time increases, after a while, the connections that must be zero going to grow, and then the whole of the learning process is destroyed.